visual-tree-search-browser-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
playwright==1.41.2
python-dotenv==1.0.1